+++ src/client/app/shared/abstract-groups-items/abstract-group.ts
